column preferences as overlay
