+++ apps/modernization-ui/src/apps/search/layout/result/header/options/table/column-preference-panel.module.scss
@@ -1,0 +1,64 @@
+@use 'styles/colors';
+@use 'styles/borders';
+
+.panel {
+    width: 14.3125rem;
+
+    display: flex;
+    flex-direction: column;
+
+    header {
+        display: flex;
+        justify-content: space-between;
+
+        padding: 1.25rem;
+
+        @extend %thin-bottom;
+
+        svg {
+            cursor: pointer;
+        }
+
+        label {
+            color: var(--base-black, #000);
+            text-align: center;
+            font-size: 1.375rem; // h5
+            font-weight: 700;
+        }
+    }
+
+    footer {
+        border-top: 1px solid var(--base-lighter, #dcdee0);
+        display: flex;
+        justify-content: flex-end;
+        padding: 1.5rem;
+    }
+}
+
+.preferences {
+    width: 14.3125rem;
+
+    .preference {
+        display: flex;
+        flex: 1 0 0;
+        align-items: center;
+        padding-left: 0.5rem;
+
+        height: 2.75rem;
+        border-bottom: 1px solid colors.$base-lighter;
+
+        .handle {
+            cursor: pointer;
+            display: flex;
+            width: 2.625rem;
+            height: 2.625rem;
+            justify-content: center;
+            align-items: center;
+            flex-shrink: 0;
+        }
+
+        .check {
+            margin: 0;
+        }
+    }
+}
