+++ apps/modernization-ui/src/overlay/overlay-panel.module.scss
@@ -1,6 +1,8 @@
 @use 'styles/borders';
 
 .overlay {
+    --modal-min-width: 400px;
+
     display: flex;
     position: relative;
     align-items: center;
@@ -10,7 +12,7 @@
         z-index: 100;
         top: 0;
 
-        min-width: 400px;
+        min-width: var(--modal-min-width);
         margin: 2.5rem 0.5rem 0 0.5rem;
         padding: 0;
